final2/backend/requirements.txt
Flask
google-cloud-aiplatform
google-cloud-storage
Pillow
python-dotenv
gunicorn
matplotlib
google-generativeai